removed useless code

+++ src/com/myne145/serverlauncher/gui/tabs/serverdashboard/charts/CPUChart.java
@@ -30,7 +30,6 @@ public class CPUChart extends JPanel {
         chart = createChart();
 
         // Create a ChartPanel to display the chart
-
         chartPanel = new XChartPanel<>(chart);
         chartPanel.setPreferredSize(new Dimension(130, 150));
         for(MouseListener mouseListener : chartPanel.getMouseListeners())
@@ -70,11 +69,9 @@ public class CPUChart extends JPanel {
 
     public void updateChartData() {
         int cpuLoad = (int) Math.ceil(systemMXBean.getCpuLoad() * 100);
-//        if (cpuLoad != 0) {
-            chart.updatePieSeries("cpu_usage", cpuLoad);
-            chart.getStyler().setSumFormat(((int) cpuLoad + "%%"));
-            chart.updatePieSeries("empty", 100 - cpuLoad);
-            repaint();  // Repaint the panel to update the chart
-//        }
+        chart.updatePieSeries("cpu_usage", cpuLoad);
+        chart.getStyler().setSumFormat(cpuLoad + "%%");
+        chart.updatePieSeries("empty", 100 - cpuLoad);
+        repaint();
     }
 }
